+++ src/rules/index.js
@@ -10,6 +10,7 @@ import blockOpeningBraceNewlineBefore from "./block-opening-brace-newline-before
 import blockOpeningBraceSpaceAfter from "./block-opening-brace-space-after"
 import blockOpeningBraceSpaceBefore from "./block-opening-brace-space-before"
 import colorHexCase from "./color-hex-case"
+import colorHexLength from "./color-hex-length"
 import colorNoInvalidHex from "./color-no-invalid-hex"
 import commentEmptyLineBefore from "./comment-empty-line-before"
 import commentSpaceInside from "./comment-space-inside"
@@ -91,6 +92,7 @@ export default {
   "block-opening-brace-space-after": blockOpeningBraceSpaceAfter,
   "block-opening-brace-space-before": blockOpeningBraceSpaceBefore,
   "color-hex-case": colorHexCase,
+  "color-hex-length": colorHexLength,
   "color-no-invalid-hex": colorNoInvalidHex,
   "comment-empty-line-before": commentEmptyLineBefore,
   "comment-space-inside": commentSpaceInside,