+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Head
 
 * Added: `color-hex-case` rule.
+* Added: `color-hex-length` rule.
 
 # 0.1.2